try some new ideas (likely keep as draft PR)
This commit reorganizes the Antithesis SDK repository structure to be
idimoatic.

- Move SDK implementation from lib/ to root directory
- Remove workspace configuration and simple/ package
- Add example programs demonstrating SDK features:
  - assert.rs: Shows usage of assertion macros
  - lifecycle.rs: Demonstrates setup and event functions
  - random.rs: Shows random number generation capabilities
- Implement AntithesisRng struct for better rand crate integration
