+++ Cargo.toml
@@ -1,3 +1,32 @@
-[workspace]
-members = ["simple", "lib"]
-resolver = "2"
+[package]
+name = "antithesis_sdk"
+version = "0.2.2"
+edition = "2021"
+rust-version = "1.62.1"
+license = "MIT"
+readme = "README.md"
+keywords = ["antithesis", "testing", "fuzzing", "quality"]
+categories= ["development-tools::testing"] # https://crates.io/category_slugs
+repository = "https://github.com/antithesishq/antithesis-sdk-rust"
+homepage = "https://github.com/antithesishq/antithesis-sdk-rust"
+documentation = "https://antithesis.com/docs/using_antithesis/sdk/rust/overview/"
+description = """
+Rust SDK for the Antithesis autonomous software testing platform.
+"""
+
+[dependencies]
+serde = { version = "1.0.113", features = ["derive"] }
+serde_json = "1.0.25"
+rand = "0.8"
+
+# needed only if full feature is set
+rustc_version_runtime = {version = "0.3", optional = true}
+once_cell = {version = "1", optional = true}
+libloading = {version = "0.8", optional = true}
+libc = {version = "0.2.64", optional = true}
+linkme = {version = "0.3.17", optional = true}
+
+
+[features]
+default = ["full"]
+full = ["dep:libloading", "dep:libc", "dep:linkme", "dep:once_cell", "dep:rustc_version_runtime"]
