--- simple/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "simple"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-antithesis_sdk = { path = "../lib", default-features = false }
-serde_json = "1.0.25"
-
-
-[features]
-default = ["antithesis_sdk/full"]
