+++ src/prelude.rs
@@ -1,17 +1,18 @@
 pub use crate::assert::{assert_raw, AssertType};
 pub use crate::assert_always;
-pub use crate::assert_always_or_unreachable;
-pub use crate::assert_reachable;
-pub use crate::assert_sometimes;
-pub use crate::assert_unreachable;
 pub use crate::assert_always_greater_than;
 pub use crate::assert_always_greater_than_or_equal_to;
 pub use crate::assert_always_less_than;
 pub use crate::assert_always_less_than_or_equal_to;
+pub use crate::assert_always_or_unreachable;
+pub use crate::assert_always_some;
+pub use crate::assert_reachable;
+pub use crate::assert_sometimes;
+pub use crate::assert_sometimes_all;
 pub use crate::assert_sometimes_greater_than;
 pub use crate::assert_sometimes_greater_than_or_equal_to;
 pub use crate::assert_sometimes_less_than;
 pub use crate::assert_sometimes_less_than_or_equal_to;
-pub use crate::assert_always_some;
-pub use crate::assert_sometimes_all;
+pub use crate::assert_unreachable;
+pub use crate::random::AntithesisRng;
 pub use crate::{antithesis_init, lifecycle, random};
