+++ tests/sdk_info.rs
