+++ tests/setup_complete_with_details.rs
