+++ tests/setup_complete_without_details.rs
